Create axiosAuth function
This function is used in places where we need to pass an authorization
header to the server to make it more ergonomic. It gets the token from
localStorage, adds it to the headers object, and then returns a function
that can be used in place of the normal `axios` function for requests.