Implement FriendDetail page
Because a user can be logged in without having first hit the `/friends`
route, we have to account for the possibility that a friend with that id
isn't loaded.

Right now the strategy is to check if state.friends.length
> 0, and if it isn't, then trigger fetchFriends. However this doesn't
account for the case where a user navigates directly to a URL of a
non-existant user ID.

Maybe the best way to ensure that is to have a cascading check is
areFriendsLoaded and then checking after that if friend is undefined. If
so, show a 'No friend with that ID' component.