Add FriendsContainer
To avoid the need of re-fetching friends over and over again, I made a
container that will fetch the friends from the server once, and then
wraps the `/friends` route. Now There are two componets the Friends and
FriendDetail that render from inside of the Friends page depending on
the route, as well as a fallback component when attempting to load a
FriendDetail page that does not contain a valid id.