Simplify the reducers and actions
There was a lot of parallel actions and cases in the reducer so I
decided to pair down. Most of our requests follow a similar pattern:
we'll send a request to the server, if successful it returns an array of
friends that we will save, if there is an error we want to take the
error and display it.

It can be loaded, successful, or error out.

The only unique request is the login, because with a successful login, I
want to have the addTokenToLocalStorage middleware intercept that token,
and save it to local storage.